backend/requirements.txt
setuptools
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt
django-cors-headers==4.3.1
dj-database-url==2.1.0
whitenoise==6.6.0
scikit-learn
numpy
